Rename config -> CFG, configuration -> config
