+++ esmvalcore/config/__init__.py
@@ -1,0 +1,5 @@
+from ._config_object import CFG
+
+__all__ = [
+    'CFG',
+]
