+++ esmvalcore/config/drs-default.yml
