+++ esmvalcore/config/drs-smhi.yml
