--- esmvalcore/configuration/__init__.py
@@ -1,5 +1,0 @@
-from ._config_object import config
-
-__all__ = [
-    'config',
-]
